Module Evm_node_lib_dev.Block_producer

type parameters = {

cctxt : Tezos_client_base.Client_context.wallet;

smart_rollup_address : string;

sequencer_key : Tezos_client_base.Client_keys.sk_uri;

maximum_number_of_chunks : int;

}

val start : parameters -> unit Tezos_base.TzPervasives.tzresult Lwt.t

start parameters starts the events follower.

val shutdown : unit -> unit Lwt.t

shutdown () stops the events follower.

val produce_block : force:bool -> timestamp:Tezos_base.TzPervasives.Time.Protocol.t -> int Tezos_base.TzPervasives.tzresult Lwt.t

produce_block ~force ~timestamp takes the transactions in the tx pool and produces a block from it, returns the number of transaction in the block. The block is not produced if the list of transactions is empty and force is set to false.

module Internal_for_tests : sig ... end