Module type Block_storage_sig.S

val current_block_number : unit -> Evm_node_lib_dev_encoding.Ethereum_types.quantity Tezos_base.TzPervasives.tzresult Lwt.t

current_block_number () returns the most recent processed and stored block number.

val nth_block : full_transaction_object:bool -> Z.t -> Evm_node_lib_dev_encoding.Ethereum_types.block Tezos_base.TzPervasives.tzresult Lwt.t

nth_block ~full_transaction_object n returns the nth processed and stored block.

If full_transaction_object is true, returns the transaction objects, the transactions hashes otherwise.

val block_by_hash : full_transaction_object:bool -> Evm_node_lib_dev_encoding.Ethereum_types.block_hash -> Evm_node_lib_dev_encoding.Ethereum_types.block Tezos_base.TzPervasives.tzresult Lwt.t

block_by_hash ~full_transaction_object hash returns the block with the given hash.

If full_transaction_object is true, returns the transaction objects, the transactions hashes otherwise.

val block_receipts : Z.t -> Evm_node_lib_dev_encoding.Transaction_receipt.t list Tezos_base.TzPervasives.tzresult Lwt.t

block_receipts n returns the receipts of the nth processed and stored block.

val transaction_receipt : Evm_node_lib_dev_encoding.Ethereum_types.hash -> Evm_node_lib_dev_encoding.Transaction_receipt.t option Tezos_base.TzPervasives.tzresult Lwt.t

transaction_receipt tx_hash returns the receipt of tx_hash.

val transaction_object : Evm_node_lib_dev_encoding.Ethereum_types.hash -> Evm_node_lib_dev_encoding.Ethereum_types.transaction_object option Tezos_base.TzPervasives.tzresult Lwt.t

transaction_object tx_hash returns the informations of tx_hash.