Module Evm_node_lib_dev.Blueprint_events

Events related to the publication of blueprints.

val section : string list

val publisher_is_ready : unit -> unit Lwt.t

publisher_is_ready () advertises that the worker is ready to accept requests.

val publisher_shutdown : unit -> unit Lwt.t

publisher_shutdown () advertises that the worker has been shutdown and will not accept requests anymore.

val blueprint_applied : Evm_node_lib_dev_encoding.Ethereum_types.block -> Tezos_base.TzPervasives.Time.System.Span.t -> unit Lwt.t

blueprint_applied block duration advertizes that a blueprint leading to block has been applied in duration time onto the local state.

val blueprint_injected : Z.t -> unit Lwt.t

blueprint_injected level advertizes that a blueprint for level level has been forwarded to a rollup node

val blueprint_injected_on_inbox : Z.t -> unit Lwt.t

blueprint_injected_on_inbox level advertizes that a blueprint for level level has been forwarded to a rollup node for injection on the shared inbox

val blueprint_injected_on_DAL : level:Z.t -> nb_chunks:int -> unit Lwt.t

blueprint_injected_on_DAL level nb_chunks advertizes that a blueprint for level level and containing nb_chunks chunks has been forwarded to a rollup node for injection on the DAL

val blueprint_injection_failed : Z.t -> Tezos_base.TzPervasives.tztrace -> unit Lwt.t

blueprint_injection_failed level trace advertizes that a blueprint could not be injected for level level.

val invalid_blueprint_produced : Z.t -> unit Lwt.t

invalid_blueprint_produced level advertizes that the sequencer has tried to produce a blueprint which does not result in the publication of a new Ethereum block.

val invalid_blueprint_applied : Z.t -> unit Lwt.t

invalid_blueprint_applied level advertizes that the observer received a blueprint for level level.

val missing_blueprints : int -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> unit Lwt.t

missing_blueprints count from to_ advertizes that a sequencer has detect it is missing count blueprints in the provided range. This means the sequencer store is inconsistent.

val catching_up : Z.t -> Z.t -> unit Lwt.t

catching_up min max advertizes that the sequencer is reinjecting blueprints from level min to max because the rollup node is lagging.

val blueprint_proposal : Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Tezos_base.TzPervasives.Time.System.Span.t -> unit Lwt.t

blueprint_proposal level duration advertizes the sequencer has crafted a blueprint for level in duration time.

val blueprint_production : Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Tezos_base.TzPervasives.Time.System.Span.t -> unit Lwt.t

blueprint_production level duration advertizes the sequencer has fully produced a blueprint for level in duration time.

val worker_request_failed : Blueprints_publisher_types.Request.view -> Tezos_base.TzPervasives.Error_monad.tztrace -> unit Lwt.t

worker_request_failed request_view state errs warns that the blueprints_publisher worker encountered errors errs.