Module Evm_node_lib_dev.Blueprints_follower

type handler = Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Blueprint_types.with_events -> [ `Check_for_reorg of Evm_node_lib_dev_encoding.Ethereum_types.quantity | `Continue ] Tezos_base.TzPervasives.tzresult Lwt.t

A handler is a function that is called for every blueprint fetched from a remote EVM endpoint.

val start : time_between_blocks:Evm_node_config.Configuration.time_between_blocks -> evm_node_endpoint:Uri.t -> next_blueprint_number:Evm_node_lib_dev_encoding.Ethereum_types.quantity -> handler -> 'a Tezos_base.TzPervasives.tzresult Lwt.t

start ~time_between_blocks ~evm_node_endpoint ~get_next_blueprint_number k is a never-returning function which iterates over blueprints streamed by evm_node_endpoint, calling k for each blueprint in order.

next_blueprint_number is the height of the current local head, while time_between_blocks is used to detect when a connection to evm_node_endpoint is stalling.