Module Evm_node_lib_dev.Blueprints_publisher

type blueprints_range = from:Evm_node_lib_dev_encoding.Ethereum_types.quantity -> to_:Evm_node_lib_dev_encoding.Ethereum_types.quantity -> (Evm_node_lib_dev_encoding.Ethereum_types.quantity * Blueprint_types.payload) list Tezos_base.TzPervasives.tzresult Lwt.t

val start : blueprints_range:blueprints_range -> rollup_node_endpoint:Uri.t -> config:Evm_node_config.Configuration.blueprints_publisher_config -> latest_level_seen:Z.t -> keep_alive:bool -> drop_duplicate:bool -> order_enabled:bool -> unit -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val shutdown : unit -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val publish : Z.t -> Blueprints_publisher_types.Request.payload -> unit Tezos_base.TzPervasives.tzresult Lwt.t

publish level payload sends a request to the publisher worker to forward the chunked blueprint payload for level level to the rollup node.

val new_rollup_block : int32 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

new_rollup_block rollup_level tells the worker that a new rollup node block.