Module Evm_node_lib_dev.Broadcast

val create_blueprint_stream : unit -> Blueprint_types.with_events Lwt_stream.t * Lwt_watcher.stopper

create_blueprint_stream () returns a new stream that can be used to be notified when a new blueprint is being successfully applied on top of the local state.

type message =

| Blueprint of Blueprint_types.with_events

Type of messages that are broadcasted to all evm nodes.

val message_encoding : message Tezos_base.TzPervasives.Data_encoding.t

val create_broadcast_stream : unit -> message Lwt_stream.t * Lwt_watcher.stopper

create_broadcast_stream () returns a new stream that can be used to be notified of all messages, including a new blueprint being successfully applied on top of the local state.

val notify : message -> unit

notify message advertizes message to every relevant streams created with create_stream or create_blueprint_stream still active.