Module Durable_storage_path.Trace

val root_indexed_by_hash : transaction_hash:path option -> path

Path under which all traces are stored when the given hash parameter is provided, otherwise it will just provide the path to the root.

val input : path

Path where is stored the input of the tracer.

val output_gas : transaction_hash:path option -> path

Path where is stored the gas consumed by the call.

val output_failed : transaction_hash:path option -> path

Path where the result of the traced transaction is stored.

val output_return_value : transaction_hash:path option -> path

Path where is stored the value returned by the transaction's execution.

val logs_length : transaction_hash:path option -> path

val opcode : transaction_hash:path option -> int -> path

val call_trace_length : transaction_hash:path option -> path

Path where is stored the number of call trace items

val call_trace : transaction_hash:path option -> int -> path

Path where is stored the ieth trace