Module Evm_node_lib_dev.Events

General purposes events.

val section : string list

Default section for events.

val received_upgrade : string -> unit Lwt.t

received_upgrade payload advertises that the sequencer received an upgrade of payload payload.

val pending_upgrade : Evm_node_lib_dev_encoding.Evm_events.Upgrade.t -> unit Lwt.t

pending_upgrade upgrade advertises that the EVM node is aware that an upgrade is pending.

val applied_upgrade : Evm_node_lib_dev_encoding.Ethereum_types.hash -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> unit Lwt.t

applied_upgrade root_hash level advertises that the kernel of the EVM node successfully upgraded to root_hash with the levelth blueprint.

val failed_upgrade : Evm_node_lib_dev_encoding.Ethereum_types.hash -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> unit Lwt.t

failed_upgrade root_hash level advertises that the kernel of the EVM node failed to upgrade to root_hash with the levelth blueprint.

val ignored_kernel_arg : unit -> unit Lwt.t

ignored_kernel_arg () advertises that the EVM node has ignored the path to the initial kernel given as a command-line argument since its EVM state was already initialized.

val catching_up_evm_event : from:int32 -> to_:int32 -> unit Lwt.t

catching_up_evm_event ~from ~to_ advertises that the sequencer is catching up on event produced by the evm kernel in the rollup node from L1 level from to to_.

val is_ready : rpc_addr:string -> rpc_port:int -> websockets:bool -> backend:Evm_node_config.Configuration.rpc_server -> unit Lwt.t

is_ready ~rpc_addr ~rpc_port ~websockets ~backend advertises that the sequencer is ready and listens to rpc_addr:rpc_port.

val private_server_is_ready : rpc_addr:string -> rpc_port:int -> websockets:bool -> backend:Evm_node_config.Configuration.rpc_server -> unit Lwt.t

private_server_is_ready ~rpc_addr ~rpc_port ~websockets ~backend advertises that the private rpc server is ready and listens to rpc_addr:rpc_port.

val rpc_server_error : exn -> unit

val shutdown_rpc_server : private_:bool -> unit Lwt.t

shutdown_rpc_server ~private_ () advertises that the RPC server was shut down, private_ tells whether it is the private server or not.

val shutdown_node : exit_status:int -> unit Lwt.t

shutdown_node ~exit_status advertises that the sequencer was shutdown, and exits with exit_status.

val callback_log : uri:string -> meth:string -> body:string -> unit Lwt.t

callback_log ~uri ~meth ~body is used as the debug event used as callback for resto to logs the requests.

type kernel_log_kind =

| Application

| Simulation

type kernel_log_level =

| Debug

| Info

| Error

| Fatal

val event_kernel_log : level:kernel_log_level -> kind:kernel_log_kind -> msg:string -> unit Lwt.t

Logs kernel log Debug.

val retrying_connect : endpoint:Uri.t -> delay:float -> unit Lwt.t

val preload_kernel : string -> unit Lwt.t

preload_kernel version advertizes the EVM node has preloaded in the module cache the kernel version.

val patched_state : string -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> unit Lwt.t

val predownload_kernel : Tezos_base.TzPervasives.Hex.t -> unit Lwt.t

predownload_kernel root_hash advertizes the EVM node has downloaded all preimages under root_hash.

val sandbox_started : Z.t -> unit Lwt.t

sandbox_started level advertizes that sandbox mode started on top of level level.

val cannot_fetch_time_between_blocks : Evm_node_config.Configuration.time_between_blocks -> Tezos_base.TzPervasives.tztrace -> unit Lwt.t

val invalid_node_da_fees : node_da_fees:Z.t -> kernel_da_fees:Z.t -> block_number:Evm_node_lib_dev_encoding.Ethereum_types.quantity option -> call:Tezos_base.TzPervasives.Data_encoding.json -> unit Lwt.t

val deprecation_note : string -> unit Lwt.t

val wasm_pvm_fallback : unit -> unit Lwt.t

wasm_pvm_fallback () advertises that the node has to fallback to the PVM to execute a block, which is slow.

val missing_chain_id : unit -> unit Lwt.t

missing_chain_id () advertises that the node could not check the consistency of the stored chain id with the selected network.

val downloading_file : ?size:int -> string -> unit Lwt.t

downloading_file ?size url advertises that the node is downloading the file at url, and explicitly mentions its size if provided.

type download_error =

| Http_error of Cohttp.Code.status_code

| Exn of exn

val download_failed : string -> download_error -> unit Lwt.t

download_failed url status advertises that the download of url failed with the given status code.

val importing_snapshot : unit -> unit Lwt.t