Module Evm_store.Blocks

val store : conn -> Evm_node_lib_dev_encoding.Ethereum_types.block -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val find_with_level : full_transaction_object:bool -> conn -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Evm_node_lib_dev_encoding.Ethereum_types.block option Tezos_base.TzPervasives.tzresult Lwt.t

find_with_level ~full_transaction_object conn level returns the block if it's present in the storage. If full_transaction_object is set to true, it will also retrieve the transactions objects part of the block.

val find_with_hash : full_transaction_object:bool -> conn -> Evm_node_lib_dev_encoding.Ethereum_types.block_hash -> Evm_node_lib_dev_encoding.Ethereum_types.block option Tezos_base.TzPervasives.tzresult Lwt.t

Same as find_with_level but finds with the block hash instead of block number.

val find_hash_of_number : conn -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Evm_node_lib_dev_encoding.Ethereum_types.block_hash option Tezos_base.TzPervasives.tzresult Lwt.t

val find_number_of_hash : conn -> Evm_node_lib_dev_encoding.Ethereum_types.block_hash -> Evm_node_lib_dev_encoding.Ethereum_types.quantity option Tezos_base.TzPervasives.tzresult Lwt.t

val clear_after : conn -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> unit Tezos_base.TzPervasives.tzresult Lwt.t