Module Evm_store.Transactions

val store : conn -> Evm_node_lib_dev_encoding.Transaction_info.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val find_receipt : conn -> Evm_node_lib_dev_encoding.Ethereum_types.hash -> Evm_node_lib_dev_encoding.Transaction_receipt.t option Tezos_base.TzPervasives.tzresult Lwt.t

val find_object : conn -> Evm_node_lib_dev_encoding.Ethereum_types.hash -> Evm_node_lib_dev_encoding.Ethereum_types.transaction_object option Tezos_base.TzPervasives.tzresult Lwt.t

val receipts_of_block_number : conn -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Evm_node_lib_dev_encoding.Transaction_receipt.t list Tezos_base.TzPervasives.tzresult Lwt.t

receipts_of_block_number conn block_number returns all the receipts found from level block_number. The function does not check if the block exists.