Module Evm_node_lib_dev.Evm_websocket

val cohttp_callback : max_message_length:int -> Rpc_encodings.websocket_handler -> Cohttp_lwt_unix.Server.conn -> Cohttp.Request.t -> 'body -> Cohttp_lwt_unix.Server.response_action Lwt.t

Callback to use for a websocket endpoint in a Cohttp server. cohttp_callback ~max_message_length handler conn req body upgrades the connection for request req to the websocket protocol and starts worker that processes incoming frames and writes in return in the websocket. max_message_length specifies the maximum size of message accepted by the server, over which the connection will be closed.

val on_conn_closed : Cohttp_lwt_unix.Server.conn -> unit

Callback to be called by Cohttp when it detects a closed connection before any read/write happens. on_conn_closed conn stops the websocket worker and cleans resources associated to connection conn, if it exists.