Module Evm_node_lib_dev.Injector

val send_raw_transaction : keep_alive:bool -> base:Uri.t -> string -> (Rpc_encodings.Send_raw_transaction.output, string) Stdlib.result Tezos_base.TzPervasives.tzresult Lwt.t

send_raw_transaction ~keep_alive ~base raw_txn uses the eth_sendRawTransaction RPC method to inject raw_txn to the endpoint base. It will retry as many time as necessary if keep_alive is set to true, or return an error otherwise.