Module Evm_node_lib_dev.Observer

val main : ?network:Evm_node_config.Configuration.supported_network -> ?kernel_path:Wasm_debugger.kernel -> data_dir:string -> config:Evm_node_config.Configuration.t -> no_sync:bool -> init_from_snapshot:bool -> unit -> unit Tezos_base.TzPervasives.tzresult Lwt.t

main ?kernel_path ~data_dir ~config ~no_sync () starts the main event-loop of the Observer, consuming the blueprints received from the EVM node endpoint, unless no_sync is true.

If init_from_snapshot is true and network is set, then the node will fetch and import the latest known snapshot for this network if data_dir was not already created.