Module Evm_node_lib_dev.Sequencer_blueprint

type t

val chunk_encoding : t Tezos_base.TzPervasives.Data_encoding.t

val chunk_to_rlp : t -> Evm_node_lib_dev_encoding.Rlp.item

chunk_to_rlp chunk encodes a chunk into its RLP format.

val prepare : cctxt:Tezos_client_base.Client_context.wallet -> sequencer_key:Tezos_client_base.Client_keys.sk_uri -> timestamp:Tezos_base.TzPervasives.Time.Protocol.t -> number:Evm_node_lib_dev_encoding.Ethereum_types.quantity -> parent_hash:Evm_node_lib_dev_encoding.Ethereum_types.block_hash -> delayed_transactions:Evm_node_lib_dev_encoding.Ethereum_types.hash list -> transactions:string list -> t list Tezos_base.TzPervasives.tzresult Lwt.t

prepare ~secret_key ~timestamp ~number ~parent_hash ~delayed_transactions ~transactions creates a sequencer blueprint at timestamp with a given number containing transactions, signed with secret_key. Returns the list of prepared chunks.

val create_inbox_payload : smart_rollup_address:string -> chunks:t list -> Blueprint_types.payload

create_inbox_payload ~smart_rollup_address ~chunks encodes the chunks into message(s) that can be read from the inbox by the kernel.

val create_dal_payloads : t list -> string list

create_dal_payloads chunks encodes the chunks into messages that can be read from DAL slots by the kernel. The chunks are prefixed by a tag.

val maximum_usable_space_in_blueprint : int -> int

maximum_usable_size_in_blueprint chunks_count returns the available space for transactions in a blueprint composed of chunks_count chunks.

val maximum_chunks_per_l1_level : int

type kernel_blueprint = {

parent_hash : Evm_node_lib_dev_encoding.Ethereum_types.block_hash;

delayed_transactions : Evm_node_lib_dev_encoding.Ethereum_types.hash list;

transactions : string list;

timestamp : Tezos_base.TzPervasives.Time.Protocol.t;

}

val kernel_blueprint_parent_hash_of_payload : Tezos_base.TzPervasives.Signature.public_key -> Blueprint_types.payload -> Evm_node_lib_dev_encoding.Ethereum_types.block_hash option

kernel_blueprint_parent_hash_of_payload sequencer bytes partially decodes fields of a kernel_blueprint and return the kernel_blueprint.parent_hash. Verify the signature of each chunk against public key sequencer.