Module Evm_node_lib_dev.Sequencer_signal

val create : cctxt:Tezos_client_base.Client_context.wallet -> sequencer_key:Tezos_client_base.Client_keys.sk_uri -> smart_rollup_address:Tezos_crypto.Hashed.Smart_rollup_address.t -> slot_ids:(int * int32) list -> [> `External of string ] Tezos_base.TzPervasives.tzresult Lwt.t

create ~cctxt ~sequencer_key ~smart_rollup_address ~slot_ids creates a DAL slot import signal for a given list slot_ids of slot_indices and their corresponding published leveles. Returns an external message input to put in the inbox.