Module type Services_backend_sig.S

module Reader : Durable_storage.READER

module Block_storage : Block_storage

val balance : Evm_node_lib_dev_encoding.Ethereum_types.address -> Evm_node_lib_dev_encoding.Ethereum_types.Block_parameter.extended -> Evm_node_lib_dev_encoding.Ethereum_types.quantity Tezos_base.TzPervasives.tzresult Lwt.t

balance address block_param returns the address's balance at block block_param.

val nonce : Evm_node_lib_dev_encoding.Ethereum_types.address -> Evm_node_lib_dev_encoding.Ethereum_types.Block_parameter.extended -> Evm_node_lib_dev_encoding.Ethereum_types.quantity option Tezos_base.TzPervasives.tzresult Lwt.t

nonce address block_param returns the address's nonce at block block_param.

val code : Evm_node_lib_dev_encoding.Ethereum_types.address -> Evm_node_lib_dev_encoding.Ethereum_types.Block_parameter.extended -> Evm_node_lib_dev_encoding.Ethereum_types.hex Tezos_base.TzPervasives.tzresult Lwt.t

code address block_param returns the address's code at block block_param.

val inject_transactions : timestamp:Tezos_base.TzPervasives.Time.Protocol.t -> smart_rollup_address:string -> transactions: (string * Evm_node_lib_dev_encoding.Ethereum_types.transaction_object) list -> Evm_node_lib_dev_encoding.Ethereum_types.hash list Tezos_base.TzPervasives.tzresult Lwt.t

inject_transactions ~timestamp ~smart_rollup_address ~transactions crafts the hashes and chunks of each transaction of transactions. Injects the chunks and returns the hashes of injected transactions.

val block_param_to_block_number : Evm_node_lib_dev_encoding.Ethereum_types.Block_parameter.extended -> Evm_node_lib_dev_encoding.Ethereum_types.quantity Tezos_base.TzPervasives.tzresult Lwt.t

val chain_id : unit -> Evm_node_lib_dev_encoding.Ethereum_types.quantity Tezos_base.TzPervasives.tzresult Lwt.t

chain_id () returns chain id defined by the rollup.

val base_fee_per_gas : unit -> Evm_node_lib_dev_encoding.Ethereum_types.quantity Tezos_base.TzPervasives.tzresult Lwt.t

base_fee_per_gas () returns base fee defined by the rollup.

val kernel_version : unit -> string Tezos_base.TzPervasives.tzresult Lwt.t

kernel_version () returns the internal kernel version (i.e the commit hash where the kernel was compiled).

val kernel_root_hash : unit -> string option Tezos_base.TzPervasives.tzresult Lwt.t

kernel_root_hash () returns the internal kernel root hash (i.e the latest root hash that was applied during an upgrade).

val simulate_call : overwrite_tick_limit:bool -> Evm_node_lib_dev_encoding.Ethereum_types.call -> Evm_node_lib_dev_encoding.Ethereum_types.Block_parameter.extended -> Evm_node_lib_dev_encoding.Ethereum_types.state_override -> Simulation.call_result Simulation.simulation_result Tezos_base.TzPervasives.tzresult Lwt.t

simulate_call call_info block_param state_override simulates a call on context block_param (optionally updated with state_override) and returns the result.

val estimate_gas : Evm_node_lib_dev_encoding.Ethereum_types.call -> Simulation.call_result Simulation.simulation_result Tezos_base.TzPervasives.tzresult Lwt.t

estimate_gas call_info asks the rollup to simulate a call, and returns the gas used to execute the call.

val storage_at : Evm_node_lib_dev_encoding.Ethereum_types.address -> Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Evm_node_lib_dev_encoding.Ethereum_types.Block_parameter.extended -> Evm_node_lib_dev_encoding.Ethereum_types.hex Tezos_base.TzPervasives.tzresult Lwt.t

storage_at address pos block_param returns the value at index pos of the account address's storage on block block_param.

val smart_rollup_address : string

val replay : Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Evm_node_lib_dev_encoding.Ethereum_types.block Tezos_base.TzPervasives.tzresult Lwt.t

val trace_transaction : Evm_node_lib_dev_encoding.Ethereum_types.hash -> Evm_node_lib_dev_encoding.Tracer_types.config -> Evm_node_lib_dev_encoding.Tracer_types.output Tezos_base.TzPervasives.tzresult Lwt.t

trace_transaction hash tracer replays the block containing the transaction hash, and traces this transaction with the specified tracer.

val coinbase : unit -> Evm_node_lib_dev_encoding.Ethereum_types.address Tezos_base.TzPervasives.tzresult Lwt.t

coinbase () returns the sequencer pool address if it exists, or the zero address.

val trace_call : Evm_node_lib_dev_encoding.Ethereum_types.call -> Evm_node_lib_dev_encoding.Ethereum_types.Block_parameter.extended -> Evm_node_lib_dev_encoding.Tracer_types.config -> Evm_node_lib_dev_encoding.Tracer_types.output Tezos_base.TzPervasives.tzresult Lwt.t