Module Evm_node_lib_dev.Simulation

type estimate_gas_input_v1 = {

call : Evm_node_lib_dev_encoding.Ethereum_types.call;

with_da_fees : bool;
(*
If true, the gas returned by the simulation include the DA gas units.
*)

}

type estimate_gas_input_v2 = {

call : Evm_node_lib_dev_encoding.Ethereum_types.call;

with_da_fees : bool;
(*
If true, the gas returned by the simulation includes the DA gas units.
*)

timestamp : Tezos_base.TzPervasives.Time.Protocol.t;
(*
Timestamp used during simulation.
*)

}

type estimate_gas_input =

| V0 of Evm_node_lib_dev_encoding.Ethereum_types.call

| V1 of estimate_gas_input_v1

| V2 of estimate_gas_input_v2

val hex_string_to_bytes : Evm_node_lib_dev_encoding.Ethereum_types.hex -> bytes

hex_string_to_bytes s transforms a hex string s into a byte string.

val of_opt : ('a -> Evm_node_lib_dev_encoding.Rlp.item) -> 'a option -> Evm_node_lib_dev_encoding.Rlp.item

val of_addr : Evm_node_lib_dev_encoding.Ethereum_types.address -> Evm_node_lib_dev_encoding.Rlp.item

val of_qty : Evm_node_lib_dev_encoding.Ethereum_types.quantity -> Evm_node_lib_dev_encoding.Rlp.item

val of_hash : Evm_node_lib_dev_encoding.Ethereum_types.hash -> Evm_node_lib_dev_encoding.Rlp.item

val rlp_v0 : Evm_node_lib_dev_encoding.Ethereum_types.call -> Evm_node_lib_dev_encoding.Rlp.item

val rlp_v1 : estimate_gas_input_v1 -> Evm_node_lib_dev_encoding.Rlp.item

val rlp_v2 : estimate_gas_input_v2 -> Evm_node_lib_dev_encoding.Rlp.item

val rlp_encode : estimate_gas_input -> bytes

Encoding used to forward the call to the kernel, to be used in simulation mode only.

type simulation_message =

| Start

| Simple of string

| NewChunked of int

| Chunk of int * string

val max_chunk_size : int

val split_in_messages : bytes -> (simulation_message list, 'a) Stdlib.result

val simulation_tag : string

Tag signaling a simulation message

val simple_tag : string

Tag signaling a simulation message containing a full simulation call

val new_chunked_tag : string

Tag signaling a simulation message starting a serie of chunks

val chunk_tag : string

Tag signaling a simulation message containing a chunk

val evaluation_tag : string

Tag indicating simulation is an evaluation

val _validation_tag : string

Tag indicating simulation is a validation

This tag can no longer be used, but if you plan to add another tag please avoid using this one to avoid mistakes.

val add_tag : string -> bytes -> bytes

add_tag tag bytes prefixes bytes by the given tag

val encode_message : simulation_message -> string

val encode : estimate_gas_input -> (string list, 'a) Stdlib.result

type execution_result = {

value : Evm_node_lib_dev_encoding.Ethereum_types.hash option;

gas_used : Evm_node_lib_dev_encoding.Ethereum_types.quantity option;

}

type call_result = (execution_result, Evm_node_lib_dev_encoding.Ethereum_types.hash) Stdlib.result

type 'a simulation_result = ('a, string) Stdlib.result

module Encodings : sig ... end

val simulation_result : bytes -> (((execution_result, Evm_node_lib_dev_encoding.Ethereum_types.hash) Stdlib.result, string) Stdlib.result, Tezos_base.TzPervasives.tztrace) Stdlib.result

val gas_estimation : bytes -> (((execution_result, Evm_node_lib_dev_encoding.Ethereum_types.hash) Stdlib.result, string) Stdlib.result, Tezos_base.TzPervasives.tztrace) Stdlib.result