Module Evm_node_lib_dev.Snapshots

type compression =

| No
(*
Produce uncompressed archive. Takes more space.
*)

| On_the_fly
(*
Compress archive on the fly. The rollup node will use less disk space to produce the snapshot but will lock the rollup node (if running) for a longer time.
*)

| After
(*
Compress archive after snapshot creation. Uses more disk space temporarily than On_the_fly but does not lock the rollup node for very long.
*)

Compression strategy for snapshot archives.

module Header : sig ... end

val export : ?snapshot_file:string -> compression:compression -> data_dir:string -> unit -> string Tezos_base.TzPervasives.tzresult Lwt.t

export ?snapshot_file ~compression ~data_dir () creates a tar gzipped archive with name snapshot_file (or a generated name in the current directory) containing a snapshot of the data of the rollup node with data directory data_dir. The path of the snapshot archive is returned.

val import : cancellable:bool -> force:bool -> data_dir:string -> snapshot_file:string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

import ~display_progress ~force ~data_dir ~snapshot_file imports the snapshot at path snapshot_file into the data directory data_dir. Import will fail if data_dir is already populated unless force is set to true. Set cancellable if you want to be able to cancel the resulting promise cleanly.

val info : snapshot_file:string -> Header.t * [ `Compressed | `Uncompressed ]

info ~snapshot_file returns information that can be used to inspect the snapshot file.