Module Evm_node_lib_dev.Threshold_encryption_types

Ocaml definitions and encoding of the datatypes used for threshold encryption. They mirror the Rust definition from `etherlink/bin_dsn_node/crates/core/src/types.rs`. Compatibility of the encodings between the Ocaml and Rust types is enforced only for (de)serialization in json format.

module Transaction : sig ... end

Module defining the transaction datatype and functions that operate on transactions.

type proposal = {

transactions : Transaction.t list;
(*
The transactions included in the proposal.
*)

delayed_transaction_hashes : Evm_node_lib_dev_encoding.Ethereum_types.hash list;
(*
The delayed transaction hashes included in the proposal.
*)

previous_block_hash : Evm_node_lib_dev_encoding.Ethereum_types.block_hash;
(*
The block hash of the last block.
*)

timestamp : Tezos_base.TzPervasives.Time.Protocol.t;
(*
The time when the proposal was crafted.
*)

current_blueprint_number : Evm_node_lib_dev_encoding.Ethereum_types.quantity;
(*
The latest blueprint number known to the EVM node.
*)

}

The type of a proposal.

val proposal_encoding : proposal Tezos_base.TzPervasives.Data_encoding.t

Encoding for proposals.

type preblock = proposal

The type of a preblock.

val preblock_encoding : preblock Tezos_base.TzPervasives.Data_encoding.t

Encoding for preblocks.

type proposal_submission_outcome =

| Tx_pool_is_locked
(*
The proposal was not submitted because the Tx_pool is locked.
*)

| Proposal_is_early
(*
The proposal was not submitted because it did not contain any transaction, and not enough time has passed since when a proposal was submitted last.
*)

| Proposal_submitted
(*
The proposal was submitted to the DSN node.
*)

Possible outcomes when submitting a proposal to the dsn node.