Module Tracer.CallTracerRead

val build_calltrace : ('a -> 'a list -> 'a) -> (int -> ('a * int) option Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a Tezos_base.TzPervasives.tzresult Lwt.t

build_calltrace end_call get_next rebuilds a call trace. end_call node list adds a list of calls to a node, and get_next i gets the data (node,depth) at line i in storage.

This is left relatively generic to facilitate testing the algorithm.