Module Wasm_runtime_callbacks.Internal_for_tests

val read_durable_value : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> (bytes, int) Stdlib.result

val store_delete : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> bool -> (Tezos_layer2_irmin_context.Irmin_context.tree, int) Stdlib.result

val store_copy : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> string -> (Tezos_layer2_irmin_context.Irmin_context.tree, int) Stdlib.result

val store_move : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> string -> (Tezos_layer2_irmin_context.Irmin_context.tree, int) Stdlib.result

val mem_tree : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> (bool, int) Stdlib.result

val store_has : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> (int, int) Stdlib.result

val store_get_hash : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> (bytes, int) Stdlib.result

val store_list_size : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> (int, int) Stdlib.result

val store_value_size : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> (int, int) Stdlib.result

val store_read : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> int -> int -> (bytes, int) Stdlib.result

val store_write : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> int -> bytes -> (Tezos_layer2_irmin_context.Irmin_context.tree * int, int) Stdlib.result

val store_write_all : Tezos_layer2_irmin_context.Irmin_context.tree -> string -> bytes -> (Tezos_layer2_irmin_context.Irmin_context.tree, int) Stdlib.result

val check_reboot_flag : Tezos_layer2_irmin_context.Irmin_context.tree -> bool * Tezos_layer2_irmin_context.Irmin_context.tree

check_reboot_flag tree returns true if the reboot flag was set in the durable storage of tree, false otherwise. Besides, it returns a new tree where the reboot flag as be removed (if it was absent, tree is returned unchanged).

module Vector : sig ... end

module Error_code : sig ... end

Error code extracted from the WASM PVM library.